Damage Values added
+ Changed health variable to a float to be used by the Slider.value.
+ Added updatedHealth(Slider healthBar), damagePlayer(int damage), and
inflictDamage(int damage). The damagePlayer is used within the
NanoByteIns.cs to change the player health and the two functions within
this function are stored in the Character.cs class. This is because we
want to inflictDamage and updateHealth for other enemies and not just
the player.

+++ GameScripts/Character.cs
@@ -3,7 +3,7 @@ using System.Collections;
 [System.Serializable]
 public class Character {
 
-	public int health;
+	public float health;
 	public float speed;
 	public Vector2 location;
 
@@ -11,18 +11,18 @@ public class Character {
 
 	public Character()
 	{
-		health = 100;
+		health = 100.0f;
 		speed = 5.0f;
 		location = Vector2.zero;
 	}
-	public Character(int aHealthVal, float aSpeedVal, Vector2 aLocation)
+	public Character(float aHealthVal, float aSpeedVal, Vector2 aLocation)
 	{
 		setHealth(aHealthVal);
 		setSpeed(aSpeedVal);
 		setLocation(aLocation);
 	}
 
-	public void setHealth(int aHealthVal)
+	public void setHealth(float aHealthVal)
 	{
 		health = aHealthVal;
 	}
@@ -35,7 +35,7 @@ public class Character {
 		location = aLocation;
 	}
  
-	public int getHealth()
+	public float getHealth()
 	{
 		return health;
 	}
